+++ FbonizziMonoGame/FbonizziMonoGame/FbonizziMonoGame.csproj
@@ -42,7 +42,7 @@
     <Compile Include="Sprites\AnimationsManager.cs" />
     <Compile Include="Assets\SplashScreenLoader.cs" />
     <Compile Include="Extensions\CollectionsExtensions.cs" />
-    <Compile Include="StringsLocalization\ILocalizedStringsRepository.cs" />
+    <Compile Include="StringsLocalization\Abstractions\ILocalizedStringsRepository.cs" />
     <Compile Include="PlatformAbstractions\ITextFileLoader.cs" />
     <Compile Include="Assets\CustomSpriteImporter.cs" />
     <Compile Include="Assets\SpriteDescription.cs" />
@@ -60,7 +60,7 @@
     <Compile Include="PlatformAbstractions\IFbonizziGame.cs" />
     <Compile Include="Input\GamePadEventArgs.cs" />
     <Compile Include="Input\GamePadListener.cs" />
-    <Compile Include="Input\IInputListener.cs" />
+    <Compile Include="Input\Abstractions\IInputListener.cs" />
     <Compile Include="Input\KeyboardEventArgs.cs" />
     <Compile Include="Input\KeyboardListener.cs" />
     <Compile Include="Input\KeyboardModifiers.cs" />
